Cleanup++

+++ src/main/java/com/kasiakab/library/repository/BorrowRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface BorrowRepository extends JpaRepository<Borrow, Integer> {
+public interface BorrowRepository extends JpaRepository<Borrow, Long> {
     List<Borrow> findByUserId(Long userId);
 
     @Query("SELECT b FROM Borrow b WHERE b.returnDate < CURRENT_DATE")
